+++ scripts/dashboard/get_data.py
@@ -1,12 +1,12 @@
 import os
 import pandas as pd
 import datetime
-import numpy as np
-import re
 
 from medperf.entities.dataset import Dataset
 from medperf import config
 from medperf.init import initialize
+
+from utils import get_institution_from_email, get_reports_path, stage_id2name
 
 from typer import Typer, Option, run
 
@@ -21,25 +21,6 @@ def get_dsets(mlcube_id):
         dset["user"] = config.comms.get_user(user_id)
 
     return dsets
-
-
-def stage_id2name(stage_str, stages_df):
-    _, code = stage_str.split()
-    code = float(code)
-    name = stages_df.loc[code, "status_name"]
-    return name
-
-
-def get_institution_from_email(email, user2institution):
-    invalid_institutions = {"gmail.com", "hotmail.com", "yahoo.com", "outlook.com"}
-    institution = user2institution.get(email, None)
-    if institution is not None:
-        return institution
-    plausible_institution = email.split("@")[1]
-    if plausible_institution in invalid_institutions:
-        return email
-    else:
-        return plausible_institution
 
 
 def build_dset_df(dsets, user2institution, stages_df):
@@ -77,13 +58,6 @@ def build_dset_df(dsets, user2institution, stages_df):
         full_table.drop(columns=["owner", "progress"], inplace=True)
 
     return full_table
-
-
-def get_reports_path(out_path, mlcube_id):
-    server_path = config.server.split("//")[1]
-    server_path = re.sub(r"[.:]", "_", server_path)
-    full_path = os.path.join(out_path, server_path, str(mlcube_id))
-    return full_path
 
 
 def write_dsets_df(dsets_df, full_path):
